refactor[#523]  : entities/auth queryKey constants 에서 api 로 이동
