+++ src/entities/auth/api/index.ts
@@ -1,1 +1,2 @@
 export * from "./getMyInfo";
+export * from "./queryKey";
