+++ src/features/setting/api/postChangeRegion.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { authQueryKey } from "@/entities/auth/constants";
+import { authQueryKey } from "@/entities/auth/api";
 import { apiClient } from "@/shared/lib";
 import { SETTING_END_POINT } from "../constants";
